Errgroup server boot and consistency changes
